modules/backend-ecr/quail-api/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# shell extension
flask-shell-ipython

# Testing
pytest==6.2.4
pytest-flask~=1.2
factory-boy==3.1.0
pdbpp==0.10.2
pytest_postgresql==2.6.*

# Lint and code style
black~=23.3
flake8==3.9.2
flake8-blind-except==0.2.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-isort==4.0.0
isort==5.8.0
pep8-naming==0.11.1

# Debug toolbar
Flask-DebugToolbar==0.11.0

# testing permissions tools
moto==4.1.*
